# src/main/java/com/whatsapp/clone/message/MessageResponse.java
package com.whatsapp.clone.message;

import lombok.*;

import java.time.LocalDateTime;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder

/*
This is a DTO that represents a message when it is sent back to the client
typically when retrieving messages.
*/
public class MessageResponse {

    private Long id;
    private String content;
    private MessageType type;
    private MessageState state;
    private String senderId;
    private String receiverId;
    private LocalDateTime createdAt;
    private byte[] media;

}