# src/main/java/com/whatsapp/clone/user/UserResponse.java
package com.whatsapp.clone.user;

import lombok.*;

import java.time.LocalDateTime;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder

// This is the DTO used to send user related data from the back to the front in the
// form of the API response.

public class UserResponse {

    private String id;
    private String firstName;
    private String lastName;
    private String email;
    private LocalDateTime lastSeen;
    private boolean isOnline;
}